run_as: Change return type of offset extraction commands to long

+++ src/common/runas.c
@@ -829,7 +829,7 @@ int run_as_rmdir_recursive(const char *path, uid_t uid, gid_t gid)
 }
 
 LTTNG_HIDDEN
-int run_as_extract_sdt_probe_offset(int fd, const char* provider,
+long run_as_extract_sdt_probe_offset(int fd, const char* provider,
 				    const char *probe_name,
 				    uid_t uid, gid_t gid)
 {
@@ -855,7 +855,7 @@ int run_as_extract_sdt_probe_offset(int fd, const char* provider,
 }
 
 LTTNG_HIDDEN
-int run_as_extract_elf_symbol_offset(int fd, const char* function,
+long run_as_extract_elf_symbol_offset(int fd, const char* function,
 				    uid_t uid, gid_t gid)
 {
 	struct run_as_data data;
